efi: earlycon: Add support for generic framebuffers and move to fbdev subsystem
Add early console support for generic linear framebuffer devices.
This driver supports probing from cmdline early parameters
or from the device-tree using information in simple-framebuffer node.
The EFI functionality should be retained in whole.
The driver was disabled on ARM because of a bug in early_ioremap
implementation on ARM.

Change-Id: I4cc8bfb540cfecc9c684705b812acdc477d35e08
Signed-off-by: Markuss Broks <markuss.broks@gmail.com>