pinctrl: Add support for MediaTek MT8163 SoC
This commit adds support for the MediaTek MT8163 System-on-Chip (SoC)
in the pinctrl-mt8163 driver. It includes the necessary Kconfig and
Makefile updates, as well as the addition of the pinctrl-mt8163.c and
pinctrl-mtk-mt8163.h files, which contain the pin control and pinctrl
definitions for the MT8163 SoC.

The MT8163 is commonly used in various ARM-based devices, and this
commit enables proper pin control functionality for this platform.

Change-Id: Ic226926cf342cf70ee5051b060f7c89b6f5e3562
Signed-off-by: Kai Jones <19639817+488315@users.noreply.github.com>